pyproject.toml
[tool.poetry]
name = "fastapi-naive-admin"
version = "v1.2.6"
description = "fastapi-naive-admin backend service"
authors = ["Cary <yaoleiguo@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = { extras = ["standard"], version = "^0.26.0" }
loguru = "^0.7.2"
pycryptodome = "^3.20.0"
python-multipart = "^0.0.6"
setuptools = "^69.0.3"
redis = "^5.0.1"
pydantic-settings = "^2.1.0"
fastapi = "^0.109.0"
cryptography = "^42.0.2"
tortoise-orm = { extras = ["asyncmy", "asyncpg"], version = "^0.20.0" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["argon2"], version = "^1.7.4" }
pyotp = "^2.9.0"
fastapi-mail = "^1.4.1"
ipy = "^1.1"
ldap3 = "^2.9.1"


[tool.poetry.group.dev.dependencies]
aerich = "^0.7.2"
eventlet = "^0.35.2"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "aerich<1.0.0,>=0.7.2",
    "eventlet<1.0.0,>=0.35.2",
]

[tool.pdm.build]
includes = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"




[project]
name = "fastapi-naive-admin"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "梁正鹏", email = "0"},
]
dependencies = [
    "loguru>=0.7.2",
    "pycryptodome>=3.20.0",
    "python-multipart>=0.0.9",
    "setuptools>=74.1.2",
    "redis>=5.0.8",
    "pydantic-settings>=2.2.1",
    "fastapi>=0.110.0",
    "cryptography<43.0.0,>=42.0.2",
    "tortoise-orm[asyncmy,asyncpg]>=0.21.6",
    "python-jose[cryptography]>=3.3.0",
    "pyotp>=2.9.0",
    "fastapi-mail>=1.4.1",
    "IPy>=1.1",
    "ldap3>=2.9.1",
    "aiosmtplib==2.0.2; python_version ~= \"3.12\"",
    "aiosqlite==0.17.0; python_version ~= \"3.12\"",
    "annotated-types==0.6.0; python_version ~= \"3.12\"",
    "anyio==4.3.0; python_version ~= \"3.12\"",
    "argon2-cffi-bindings==21.2.0; python_version ~= \"3.12\"",
    "argon2-cffi==23.1.0; python_version ~= \"3.12\"",
    "async-timeout==4.0.3; python_version ~= \"3.12\"",
    "asyncmy>=0.2.9",
    "asyncpg==0.29.0; python_version ~= \"3.12\"",
    "blinker==1.7.0; python_version ~= \"3.12\"",
    "cffi==1.16.0; python_version ~= \"3.12\"",
    "click==8.1.7; python_version ~= \"3.12\"",
    "colorama==0.4.6; sys_platform == \"win32\" and python_version ~= \"3.12\" or platform_system == \"Windows\" and python_version ~= \"3.12\"",
    "cryptography==42.0.5; python_version ~= \"3.12\"",
    "dnspython==2.6.1; python_version ~= \"3.12\"",
    "ecdsa==0.18.0; python_version ~= \"3.12\"",
    "email-validator==2.1.1; python_version ~= \"3.12\"",
    "fastapi-mail==1.4.1; python_version ~= \"3.12\"",
    "fastapi==0.109.2; python_version ~= \"3.12\"",
    "h11==0.14.0; python_version ~= \"3.12\"",
    "httptools==0.6.1; python_version ~= \"3.12\"",
    "idna==3.6; python_version ~= \"3.12\"",
    "ipy==1.01; python_version ~= \"3.12\"",
    "iso8601==1.1.0; python_version ~= \"3.12\"",
    "ldap3==2.9.1; python_version ~= \"3.12\"",
    "loguru==0.7.2; python_version ~= \"3.12\"",
    "markupsafe==2.1.5; python_version ~= \"3.12\"",
    "pyasn1==0.5.1; python_version ~= \"3.12\"",
    "pycparser==2.21; python_version ~= \"3.12\"",
    "pycryptodome==3.20.0; python_version ~= \"3.12\"",
    "pydantic-core==2.16.3; python_version ~= \"3.12\"",
    "pydantic-settings==2.2.1; python_version ~= \"3.12\"",
    "pydantic==2.6.3; python_version ~= \"3.12\"",
    "pyotp==2.9.0; python_version ~= \"3.12\"",
    "pypika-tortoise==0.1.6; python_version ~= \"3.12\"",
    "python-dotenv==1.0.1; python_version ~= \"3.12\"",
    "python-jose[cryptography]==3.3.0; python_version ~= \"3.12\"",
    "pytz==2024.1; python_version ~= \"3.12\"",
    "pyyaml>=6.0.2",
    "redis==5.0.2; python_version ~= \"3.12\"",
    "rsa==4.9; python_version ~= \"3.12\"",
    "setuptools==69.1.1; python_version ~= \"3.12\"",
    "six==1.16.0; python_version ~= \"3.12\"",
    "sniffio==1.3.1; python_version ~= \"3.12\"",
    "starlette==0.36.3; python_version ~= \"3.12\"",
    "tortoise-orm[asyncmy,asyncpg]==0.20.0; python_version ~= \"3.12\"",
    "typing-extensions==4.10.0; python_version ~= \"3.12\"",
    "uvloop==0.19.0; (sys_platform != \"win32\" and sys_platform != \"cygwin\") and platform_python_implementation != \"PyPy\" and python_version ~= \"3.12\"",
    "watchfiles==0.21.0; python_version ~= \"3.12\"",
    "websockets==12.0; python_version ~= \"3.12\"",
    "win32-setctime==1.1.0; python_version ~= \"3.12\" and sys_platform == \"win32\"",
    "exceptiongroup>=1.2.2",
    "passlib[argon2]>=1.7.4",
    "pillow>=11.0.0",
    "asyncio>=3.4.3",
    "httpx==0.28.0",
    "requests>=2.32.3",
    "nicegui==2.9.1",
    "opencv-python>=4.11.0.86",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
package-mode = false

[tool.pdm.scripts]
dev = "python3 main.py"